requirements.txt
torch
torchvision
pillow
matplotlib
evaluate
tqdm
mlflow-skinny
schedulefree
torchmetrics
python-dotenv
boto3